big cleanup in PyBtcWallet.py, cause it was hard to read refactor DlgRemoveWallet and DlgExportWO out of qtdialogs.py move OpenPaperBackupDialog to DlgBackupCenter.py use named constants in DlgRestore & DlgReplaceWallet plain remove wallet GUI support
