+++ armoryengine/PyBtcWallet.py
@@ -47,24 +47,6 @@ PYROOTPKCCVER = 1 # Current version of root pub key/chain code backup format
 PYROOTPKCCVERMASK = 0x7F
 PYROOTPKCCSIGNMASK = 0x80
 
-# Only works on PyBtcWallet
-# If first arg is not PyBtcWallet call the function as if it was
-# not decorated, it should throw whatever error or do whatever it would
-# do withouth this decorator. This decorator does nothing if applied to 
-# the methods of any other class
-def CheckWalletRegistration(func):
-   def inner(*args, **kwargs):
-      if len(args)>0 and isinstance(args[0],PyBtcWallet):
-         if args[0].isRegistered():
-            return func(*args, **kwargs)
-         elif 'doRegister' in kwargs and kwargs['doRegister'] == False:
-            return func(*args, **kwargs)
-         else:
-            raise WalletUnregisteredError
-      else:
-         return func(*args, **kwargs)
-   return inner
-
 def computeSettingsId(wltId, accId):
    #armory's py implementation of hmac mangles the 2nd half but
    #we don't care for the purpose of concatenating ids
@@ -75,8 +57,8 @@ def computeSettingsId(wltId, accId):
 ################################################################################
 class PyBtcWallet(object):
    """
-   ***NOTE: This an abridged version of the previous description. Not much of
-      what's mentionned here has any relevance to the wallet design anymore,
+   ***NOTE: This an abridged version of the initial description. Not much of
+      what's mentionned here has any relevance to the wallet structure anymore,
       it is preserved for historical purposes only.
 
    I have ONLY implemented deterministic wallets, using ECDSA
@@ -120,17 +102,16 @@ class PyBtcWallet(object):
       self.addrByString = {}
       self.txAddrMap = {}    # cache for getting tx-labels based on addr search
       if USE_TESTNET or USE_REGTEST:
-         self.addrPoolSize = 10  # this makes debugging so much easier!
+         self.addrPoolSize = 10
       else:
          self.addrPoolSize = CLI_OPTIONS.keypool
 
       # For file sync features
       self.walletPath = ''
-      self.doBlockchainSync = BLOCKCHAIN_READONLY
       self.lastSyncBlockNum = 0
 
       # Private key encryption details
-      self.useEncryption  = False
+      self.useEncryption = False
       self.testedComputeTime = None
       self.kdfMemoryReq = 0
 
@@ -142,7 +123,6 @@ class PyBtcWallet(object):
       self._dbId        = None
       self._settingsId  = None
 
-      self.lastComputedChainAddr160 = None
       self.lastComputedChainIndex   = 0
       self._highestUsedChainIndex    = -1
 
@@ -163,36 +143,178 @@ class PyBtcWallet(object):
          self.loadFromProto(proto)
 
    #############################################################################
-   def isWltSigningAnyLockbox(self, lockboxList):
-      for lockbox in lockboxList:
-         for addr160 in lockbox.a160List:
-            if addr160 in self.addrMap:
-               return True
-      return False
+   ## setup routines
+   @staticmethod
+   def loadFromBridge(walletId):
+      wallet = PyBtcWallet(wltId=walletId)
+      walletProto = wallet.bridgeWalletObj.getData()
+      wallet.loadFromProto(walletProto)
+      return wallet
+
+   def loadFromProto(self, payload):
+      self._walletId       = payload.walletId
+      self._accountId      = payload.accountId
+      self._dbId           = payload.dbId
+      self._settingsId     = computeSettingsId(self._walletId, self._accountId)
+      self.bridgeWalletObj = BridgeWalletWrapper(self._walletId, self._accountId)
+
+      self.labelName   = payload.label
+      self.labelDescr  = payload.desc
+
+      self.useEncryption = payload.usesEncryption
+      self.lastComputedChainIndex = payload.lookupCount
+      self._highestUsedChainIndex = payload.useCount
+      self.watchingOnly = payload.watchingOnly
+      self.addressTypes = payload.addressTypes
+      self.defaultAddressType = payload.defaultAddressType
+      self.kdfMemoryReq = payload.kdfMemReq * (1024**2)
+
+      #addrMap and chainIndexMap
+      for addr in payload.addressData:
+         addrObj = PyBtcAddress(self)
+         addrObj.loadFromProto(addr)
+         self.addAddress(addrObj)
+
+      #comments
+      for commentIt in payload.comments:
+         self.commentsMap[commentIt.key] = commentIt.val
+
+   ####
+   def register(self, isNew):
+      TheBridge.service.registerWallet(self.walletId, self.accountId, isNew)
+
+   ####
+   def syncData(self):
+      dataProto = self.bridgeWalletObj.getData()
+      self.lastComputedChainIndex = dataProto.lookupCount
+      self._highestUsedChainIndex = dataProto.useCount
+
+      for addr in dataProto.addressData:
+         addrObj = PyBtcAddress(self)
+         addrObj.loadFromProto(addr)
+         self.addAddress(addrObj, updateOnly=True)
+
+   def updateBalancesAndCount(self):
+      result = self.bridgeWalletObj.getBalanceAndCount()
+      self.balance = Balances.fromProto(result)
+
+   ####
+   def addAddress(self, addrObj, updateOnly: bool=False):
+      prefixedHash = addrObj.getPrefixedAddr()
+      if prefixedHash in self.addrMap and updateOnly:
+         return
+
+      if addrObj.parentWallet == None:
+         addrObj.parentWallet = self
+      self.addrMap[prefixedHash] = addrObj
+      self.linearAddr160List.append(prefixedHash)
+
+      self._highestUsedChainIndex = max(
+         addrObj.chainIndex,
+         self._highestUsedChainIndex)
+      self.addrByString[addrObj.getAddressString()] = prefixedHash
+
+   def getAddrDataFromDB(self):
+      result = self.bridgeWalletObj.getAddrCombinedList()
+
+      #update addr map
+      for addrProto in result.updatedAssets:
+         addrObj = PyBtcAddress()
+         addrObj.loadFromProto(addrProto)
+         self.addAddress(addrObj)
+
+      #update balances and txio count
+      for addrCombinedData in result.balances:
+         addr = addrCombinedData.scrAddr
+         if addr in self.addrMap:
+            addrObj = self.addrMap[addr]
+            addrObj.balance = Balances.fromProto(addrCombinedData.balances)
+         else:
+            print ("[getAddrDataFromDB] missing address " + addr.hex())
 
    #############################################################################
-   def getWalletVersion(self):
-      return (getVersionInt(self.version), getVersionString(self.version))
-
-   #############################################################################
-   def getTimeRangeForAddress(self, addr160):
-      if addr160 not in self.addrMap:
-         return None
+   ## comments
+   def getComment(self, hashVal):
+      """
+      This method is used for both address comments, as well as tx comments
+      In the first case, use the 20-byte addr160. Use 32-byte tx
+      hash for the tx-comment case.
+      """
+      if hashVal in self.commentsMap:
+         return self.commentsMap[hashVal]
       else:
-         return self.addrMap[addr160].getTimeRange()
+         return ''
 
-   #############################################################################
-   def getBlockRangeForAddress(self, addr160):
-      if addr160 not in self.addrMap:
-         return None
+   def setComment(self, hashVal, newComment):
+      """
+      This method is used for both address comments, as well as tx comments
+      In the first case, use the 20-byte addr160. Use 32-byte tx
+      hash for the tx-comment case.
+      """
+      self.commentsMap[hashVal] = newComment
+      self.bridgeWalletObj.setComment(hashVal, newComment)
+
+   ####
+   def getAddrCommentIfAvail(self, txHash):
+      # If we haven't extracted relevant addresses for this tx, yet -- do it
+      if txHash not in self.txAddrMap:
+         self.txAddrMap[txHash] = []
+         try:
+            tx = TheBDM.bdv().getTxByHash(txHash)
+         except:
+            return ''
+         if tx.isInitialized():
+            for i in range(tx.getNumTxOut()):
+               txout = tx.getTxOutCopy(i)
+               stype = getTxOutScriptType(txout.getScript())
+               scrAddr = tx.getScrAddrForTxOut(i)
+
+               if stype in CPP_TXOUT_HAS_ADDRSTR:
+                  addrStr = scrAddr_to_addrStr(scrAddr)
+                  addr160 = addrStr_to_hash160(addrStr)[1]
+                  if self.hasAddr(addr160):
+                     self.txAddrMap[txHash].append(addr160)
+               else:
+                  pass
+
+      addrComments = []
+      for a160 in self.txAddrMap[txHash]:
+         h160 = a160[1:]
+         if h160 in self.commentsMap and '[[' not in self.commentsMap[h160]:
+            addrComments.append(self.commentsMap[h160])
+
+      return '; '.join(addrComments)
+
+   ####
+   def getAddrCommentFromLe(self, le):
+      # If we haven't extracted relevant addresses for this tx, yet -- do it
+      txHash = le.txHash
+      if txHash not in self.txAddrMap:
+         self.txAddrMap[txHash] = le.scrAddrs
+
+      addrComments = []
+      for a160 in self.txAddrMap[txHash]:
+         hash160 = a160[1:]
+         if hash160 in self.commentsMap and '[[' not in self.commentsMap[hash160]:
+            addrComments.append(self.commentsMap[hash160])
+
+      return '; '.join(addrComments)
+
+   def getCommentForLE(self, le):
+      # Smart comments for LedgerEntry objects:  get any direct comments ...
+      # if none, then grab the one for any associated addresses.
+      txHash = le.txHash
+      if txHash in self.commentsMap:
+         comment = self.commentsMap[txHash]
       else:
-         return self.addrMap[addr160].getBlockRange()
+         # [[ COMMENTS ]] are not meant to be displayed on main ledger
+         comment = self.getAddrCommentFromLe(le)
+         if comment.startswith('[[') and comment.endswith(']]'):
+            comment = ''
 
-   #############################################################################
-   def setBlockchainSyncFlag(self, syncYes=True):
-      self.doBlockchainSync = syncYes
+      return comment
 
-   #############################################################################
+   ####
    def getCommentForAddrBookEntry(self, abe):
       comment = self.getComment(abe.getAddr160())
       if len(comment)>0:
@@ -204,7 +326,7 @@ class PyBtcWallet(object):
             return comment
       return ''
 
-   #############################################################################
+   ####
    def getCommentForTxList(self, a160, txhashList):
       comment = self.getComment(a160)
       if len(comment)>0:
@@ -217,34 +339,14 @@ class PyBtcWallet(object):
       return ''
 
    #############################################################################
-   def hasAnyImported(self):
-      for a160,addr in self.addrMap.items():
-         if addr.chainIndex == -2:
-            return True
-      return False
-
-   def isRegistered(self):
-      return True
-
-   #############################################################################
-   # The IGNOREZC args on the get*Balance calls determine whether unconfirmed
-   # change (sent-to-self) will be considered spendable or unconfirmed.  This
-   # was added after the malleability issues cropped up in Feb 2014.  Zero-conf
-   # change was always deprioritized, but using --nospendzeroconfchange makes
-   # it totally unspendable
+   ## balances & tx count getters
    def getBalance(self, balType="Spendable"):
       return self.balance.getBalance(balType)
 
-   #############################################################################
    def getTxnCount(self):
       return self.balance.txCount
 
-   #############################################################################  
-   def updateBalancesAndCount(self):
-      result = self.bridgeWalletObj.getBalanceAndCount()
-      self.balance = Balances.fromProto(result)
-
-   #############################################################################
+   ####
    def getAddrBalance(self, addrHash, balType="Spendable"):
       if not self.hasAddrHash(addrHash):
          return -1
@@ -252,94 +354,132 @@ class PyBtcWallet(object):
          addrObj = self.addrMap[addrHash]
          return addrObj.getBalance(balType)
 
-   #############################################################################
-   @CheckWalletRegistration
-   def getTxLedger(self, ledgType='Full'):
-      """
-      Gets the ledger entries for the entire wallet, from C++/SWIG data structs
-      """
-      ledgBlkChain = self.getHistoryPage(0)
-      ledg = []
-      ledg.extend(ledgBlkChain)
-      return ledg
+   def getAddrTotalTxnCount(self, addrHash):
+      try:
+         addrObj = self.addrMap[addrHash]
+         return addrObj.getTxioCount()
+      except:
+         return 0
 
    #############################################################################
-   @CheckWalletRegistration
+   ## UTXO/coin selection
+   def initCoinSelectionInstance(self, height):
+      return self.bridgeWalletObj.initCoinSelectionInstance(height)
+
+   ####
    def getFullUTXOList(self):
-      """
-      DO NOT USE THIS CALL UNLESS NECESSARY.
-      This call returns *ALL* of the wallet's UTXOs. If your intent is to get
-      UTXOs to spend coins, use getUTXOListForSpendVal and pass the amount you 
-      want to spend as the argument.
-      
-      If you want to get UTXOs for browsing the history, use 
-      getUTXOListForBlockRange with the top and bottom block of the desired range
-      """
+      #return all mined UTXOs
+      from armoryengine.CoinSelection import PyUnspentTxOut
+      utxos = self.bridgeWalletObj.getUtxos(value=2**64 - 1)
+      utxoList = []
+      for utxo in utxos:
+         utxoList.append(PyUnspentTxOut().createFromBridgeUtxo(utxo))
+      return utxoList
 
-      #return full set of unspent TxOuts
-      if not self.doBlockchainSync==BLOCKCHAIN_DONOTUSE:
-         #calling this with no value argument will return the full UTXO list
-         from armoryengine.CoinSelection import PyUnspentTxOut
-         utxos = self.bridgeWalletObj.getUtxos(value=2**64 - 1)
-         utxoList = []
-         for utxo in utxos:
-            utxoList.append(PyUnspentTxOut().createFromBridgeUtxo(utxo))
-         return utxoList
-      else:
-         LOGERROR('***Blockchain is not available for accessing wallet-tx data')
-         return []
-
-   #############################################################################
-   @CheckWalletRegistration
    def getZCUTXOList(self):
-      #return full set of unspent ZC outputs
-      if not self.doBlockchainSync==BLOCKCHAIN_DONOTUSE:
-         from armoryengine.CoinSelection import PyUnspentTxOut
-         utxos = self.bridgeWalletObj.getUtxos(zc=True)
-         utxoList = []
-         for utxo in utxos:
-            utxoList.append(PyUnspentTxOut().createFromBridgeUtxo(utxo))
-         return utxoList
-      else:
-         LOGERROR('***Blockchain is not available for accessing wallet-tx data')
-         return []
+      #return all ZC UTXOs
+      from armoryengine.CoinSelection import PyUnspentTxOut
+      utxos = self.bridgeWalletObj.getUtxos(zc=True)
+      utxoList = []
+      for utxo in utxos:
+         utxoList.append(PyUnspentTxOut().createFromBridgeUtxo(utxo))
+      return utxoList
 
-   #############################################################################
-   @CheckWalletRegistration
    def getRBFTxOutList(self):
-      #return full set of unspent ZC outputs
-      if not self.doBlockchainSync==BLOCKCHAIN_DONOTUSE:
-         from armoryengine.CoinSelection import PyUnspentTxOut
-         utxos = self.bridgeWalletObj.getUtxos(rbf=True)
-         utxoList = []
-         for utxo in utxos:
-            utxoList.append(PyUnspentTxOut().createFromBridgeUtxo(utxo))
-         return utxoList
-      else:
-         LOGERROR('***Blockchain is not available for accessing wallet-tx data')
-         return []
+      #return all replaceable ZC outputs
+      from armoryengine.CoinSelection import PyUnspentTxOut
+      utxos = self.bridgeWalletObj.getUtxos(rbf=True)
+      utxoList = []
+      for utxo in utxos:
+         utxoList.append(PyUnspentTxOut().createFromBridgeUtxo(utxo))
+      return utxoList
 
    #############################################################################
+   ## address lookups
    def getAddrByHash(self, addrHash):
       if addrHash not in self.addrMap:
          return None
       return self.addrMap[addrHash]
 
-   #############################################################################
    def hasAddrHash(self, addrHash):
       return addrHash in self.addrMap
 
-   #############################################################################
    def getAddrByString(self, addrStr):
       if addrStr not in self.addrByString:
          return None
       return self.addrMap[self.addrByString[addrStr]]
 
-   #############################################################################
    def hasAddrString(self, addrStr):
       return addrStr in self.addrByString
 
+   ####
+   def getTimeRangeForAddress(self, addr160):
+      if addr160 not in self.addrMap:
+         return None
+      else:
+         return self.addrMap[addr160].getTimeRange()
+
+   def getBlockRangeForAddress(self, addr160):
+      if addr160 not in self.addrMap:
+         return None
+      else:
+         return self.addrMap[addr160].getBlockRange()
+
    #############################################################################
+   ## getting new addresses
+   def peekChangeAddr(self, addrType=AddressEntryType_Default):
+      newAddrProto = self.bridgeWalletObj.getNewAddress(addrType)
+      newAddrObj = PyBtcAddress()
+      newAddrObj.loadFromProto(newAddrProto)
+      return newAddrObj
+
+   def getNewChangeAddr(self, addrType=AddressEntryType_Default):
+      newAddrProto = self.bridgeWalletObj.getNewAddress(addrType)
+      newAddrObj = PyBtcAddress()
+      newAddrObj.loadFromProto(newAddrProto)
+      self.addAddress(newAddrObj)
+      return newAddrObj
+
+   def getNextUnusedAddress(self, addrType=AddressEntryType_Default):
+      newAddrProto = self.bridgeWalletObj.getNewAddress(addrType)
+      newAddrObj = PyBtcAddress()
+      newAddrObj.loadFromProto(newAddrProto)
+      self.addAddress(newAddrObj)
+      return newAddrObj
+
+   ####
+   def fillAddressPool(self, numPool, progressId, callback=None):
+      """
+      This is only ever called by explicit user request. We therefor assume
+      the addresses are being restored (flagged as not new)
+      """
+      self.bridgeWalletObj.extendAddressPool(
+         numPool, False, progressId, callback)
+
+   ####
+   def getAddressTypes(self):
+      return self.addressTypes
+
+   def getDefaultAddressType(self):
+      return self.defaultAddressType
+
+   def getDefaultChangeType(self):
+      #TODO: add entry in cpp wallets for default change type
+      return self.defaultChangeType
+
+   def setAddressTypeFor(self, addrObj, addrType):
+      if addrObj.addrType == addrType:
+         return
+
+      if addrType not in self.getAddressTypes():
+         raise Exception("ineligible address type")
+
+      protoAddr = self.bridgeWalletObj.setAddressTypeFor(
+         addrObj.assetId, addrType)
+      addrObj.loadFromProto(protoAddr)
+
+   #############################################################################
+   ## create/export
    @staticmethod
    def createNewWallet(replyCallback: callable, callbackId: str,
       shortLabel: str='', longLabel: str='', extraEntropy: bytes=None):
@@ -378,80 +518,7 @@ class PyBtcWallet(object):
          callbackId, replyCallback
       )
 
-   #############################################################################
-   @staticmethod
-   def loadFromBridge(walletId):
-      wallet = PyBtcWallet(wltId=walletId)
-      walletProto = wallet.bridgeWalletObj.getData()
-      wallet.loadFromProto(walletProto)
-      return wallet
-
-   def syncData(self):
-      dataProto = self.bridgeWalletObj.getData()
-      self.lastComputedChainIndex = dataProto.lookupCount
-      self._highestUsedChainIndex = dataProto.useCount
-
-      for addr in dataProto.addressData:
-         addrObj = PyBtcAddress(self)
-         addrObj.loadFromProto(addr)
-         self.addAddress(addrObj, updateOnly=True)
-
-   #############################################################################
-   def addAddress(self, addrObj, updateOnly: bool=False):
-      prefixedHash = addrObj.getPrefixedAddr()
-      if prefixedHash in self.addrMap and updateOnly:
-         return
-
-      if addrObj.parentWallet == None:
-         addrObj.parentWallet = self
-      self.addrMap[prefixedHash] = addrObj
-      self.linearAddr160List.append(prefixedHash)
-
-      self._highestUsedChainIndex = max(
-         addrObj.chainIndex,
-         self._highestUsedChainIndex)
-      self.addrByString[addrObj.getAddressString()] = prefixedHash
-
-   #############################################################################
-   def peekChangeAddr(self, addrType=AddressEntryType_Default):
-      newAddrProto = self.bridgeWalletObj.getNewAddress(addrType)
-      newAddrObj = PyBtcAddress()
-      newAddrObj.loadFromProto(newAddrProto)
-      return newAddrObj
-
-   #############################################################################
-   def getNewChangeAddr(self, addrType=AddressEntryType_Default):
-      newAddrProto = self.bridgeWalletObj.getNewAddress(addrType)
-      newAddrObj = PyBtcAddress()
-      newAddrObj.loadFromProto(newAddrProto)
-      self.addAddress(newAddrObj)
-      return newAddrObj
-
-   #############################################################################
-   def getNextUnusedAddress(self, addrType=AddressEntryType_Default):
-      newAddrProto = self.bridgeWalletObj.getNewAddress(addrType)
-      newAddrObj = PyBtcAddress()
-      newAddrObj.loadFromProto(newAddrProto)
-      self.addAddress(newAddrObj)
-      return newAddrObj
-
-   #############################################################################
-   def getHighestUsedIndex(self):
-      """
-      This only retrieves the stored value, but it may not be correct if,
-      for instance, the wallet was just imported but has been used before.
-      """
-      return self._highestUsedChainIndex + 1
-
-   #############################################################################
-   def getHighestComputedIndex(self):
-      """
-      This only retrieves the stored value, but it may not be correct if,
-      for instance, the wallet was just imported but has been used before.
-      """
-      return self.lastComputedChainIndex
-
-   #############################################################################
+   ####
    def getRootPKCC(self, pkIsCompressed=False):
       '''Get the root public key and chain code for this wallet. The key may be
          compressed or uncompressed.'''
@@ -477,7 +544,6 @@ class PyBtcWallet(object):
 
       return (wltRootPubKey, wltChainCode)
 
-   #############################################################################
    def getRootPKCCBackupData(self, pkIsCompressed=True, et16=True):
       '''
       Get the root public key and chain code for this wallet. The root pub
@@ -527,7 +593,6 @@ class PyBtcWallet(object):
       # Return the root ID & the PK/CC data.
       return (wltRootIDConcat, pkccLines)
 
-   #############################################################################
    def writePKCCFile(self, newPath):
       '''Make a copy of this wallet with only the public key and chain code.'''
       # Open the PKCC file for writing.
@@ -547,7 +612,7 @@ class PyBtcWallet(object):
       # Clean everything up.
       newFile.close()
 
-   #############################################################################
+   ####
    def forkOnlineWallet(self, newWalletFile, shortLabel='', longLabel=''):
       """
       Make a copy of this wallet that contains no private key data
@@ -587,173 +652,19 @@ class PyBtcWallet(object):
 
       onlineWallet.uniqueIDBin = self.uniqueIDBin
       onlineWallet._highestUsedChainIndex     = self._highestUsedChainIndex
-      onlineWallet.lastComputedChainAddr160  = self.lastComputedChainAddr160
       onlineWallet.lastComputedChainIndex    = self.lastComputedChainIndex
 
       onlineWallet.writeFreshWalletFile(newWalletFile, shortLabel, longLabel)
       return onlineWallet
 
-   #############################################################################
-   def testKdfComputeTime(self, callback):
-      """
-      Experimentally determines the compute time required by this computer
-      to execute with the current key-derivation parameters.  This may be
-      useful for when you transfer a wallet to a new computer that has
-      different speed/memory characteristic.
-      """
-      def handleReply(reply):
-         if reply.success == False:
-            LOGERROR("failed to test KDF")
-         else:
-            self.testedComputeTime = reply.wallet.getUnlockTime
-         callback(reply.success, self.testedComputeTime)
-      self.bridgeWalletObj.getUnlockTime(handleReply)
+   ####
+   def createBackupString(self, callback,
+      passphrase: str=None, unlockHandler: callable=None):
+      return self.bridgeWalletObj.createBackupStringForWallet(callback,
+         passphrase, unlockHandler)
 
    #############################################################################
-   def changePassphrase(self, isPriv, callbackId, callback):
-      self.bridgeWalletObj.changePassphrase(isPriv, callbackId, callback)
-
-   #############################################################################
-   def getWalletPath(self, nameSuffix=None):
-      raise Exception("fixme")
-      fpath = self.walletPath
-
-      if self.walletPath=='':
-         fpath = os.path.join(ARMORY_HOME_DIR, buildWltFileName(self.uniqueIDB58))
-
-      if not nameSuffix==None:
-         pieces = os.path.splitext(fpath)
-         if not pieces[0].endswith('_'):
-            fpath = pieces[0] + '_' + nameSuffix + pieces[1]
-         else:
-            fpath = pieces[0] + nameSuffix + pieces[1]
-      return fpath
-
-   #############################################################################
-   def getDisplayStr(self, pref="Wallet: "):
-      return '%s"%s" (%s)' % (pref, self.labelName, self.walletId)
-
-   #############################################################################
-   def getComment(self, hashVal):
-      """
-      This method is used for both address comments, as well as tx comments
-      In the first case, use the 20-byte addr160. Use 32-byte tx
-      hash for the tx-comment case.
-      """
-      if hashVal in self.commentsMap:
-         return self.commentsMap[hashVal]
-      else:
-         return ''
-
-   #############################################################################
-   def setComment(self, hashVal, newComment):
-      """
-      This method is used for both address comments, as well as tx comments
-      In the first case, use the 20-byte addr160. Use 32-byte tx
-      hash for the tx-comment case.
-      """
-      self.commentsMap[hashVal] = newComment
-      self.bridgeWalletObj.setComment(hashVal, newComment)
-
-   #############################################################################
-   def getAddrCommentIfAvail(self, txHash):
-      # If we haven't extracted relevant addresses for this tx, yet -- do it
-      if txHash not in self.txAddrMap:
-         self.txAddrMap[txHash] = []
-         try:
-            tx = TheBDM.bdv().getTxByHash(txHash)
-         except:
-            return ''
-         if tx.isInitialized():
-            for i in range(tx.getNumTxOut()):
-               txout = tx.getTxOutCopy(i)
-               stype = getTxOutScriptType(txout.getScript())
-               scrAddr = tx.getScrAddrForTxOut(i)
-
-               if stype in CPP_TXOUT_HAS_ADDRSTR:
-                  addrStr = scrAddr_to_addrStr(scrAddr)
-                  addr160 = addrStr_to_hash160(addrStr)[1]
-                  if self.hasAddr(addr160):
-                     self.txAddrMap[txHash].append(addr160)
-               else:
-                  pass
-                  #LOGERROR("Unrecognized scraddr: " + binary_to_hex(scrAddr))
-
-      addrComments = []
-      for a160 in self.txAddrMap[txHash]:
-         h160 = a160[1:]
-         if h160 in self.commentsMap and '[[' not in self.commentsMap[h160]:
-            addrComments.append(self.commentsMap[h160])
-
-      return '; '.join(addrComments)
-
-   #############################################################################
-   def getAddrCommentFromLe(self, le):
-      # If we haven't extracted relevant addresses for this tx, yet -- do it
-      txHash = le.txHash
-      if txHash not in self.txAddrMap:
-         self.txAddrMap[txHash] = le.scrAddrs
-
-      addrComments = []
-      for a160 in self.txAddrMap[txHash]:
-         hash160 = a160[1:]
-         if hash160 in self.commentsMap and '[[' not in self.commentsMap[hash160]:
-            addrComments.append(self.commentsMap[hash160])
-
-      return '; '.join(addrComments)
-
-   #############################################################################
-   def getCommentForLE(self, le):
-      # Smart comments for LedgerEntry objects:  get any direct comments ...
-      # if none, then grab the one for any associated addresses.
-      txHash = le.txHash
-      if txHash in self.commentsMap:
-         comment = self.commentsMap[txHash]
-      else:
-         # [[ COMMENTS ]] are not meant to be displayed on main ledger
-         comment = self.getAddrCommentFromLe(le)
-         if comment.startswith('[[') and comment.endswith(']]'):
-            comment = ''
-
-      return comment
-
-   #############################################################################
-   def setLabels(self, lshort, llong=''):
-      self.labelName = lshort
-      self.labelDescr = llong
-      self.bridgeWalletObj.setLabels(self.labelName, self.labelDescr)
-
-   #############################################################################  
-   def importExternalAddressBatch(self, privKeyList):
-      addr160List = []
-      for key, a160 in privKeyList:
-         self.importExternalAddressData(key)
-         addr160List.append(Hash160ToScrAddr(a160))
-      return addr160List
-
-   #############################################################################
-   def getAddrListSortedByChainIndex(self, withRoot=False):
-      """ Returns Addr160 list """
-      addrList = []
-      for addr160 in self.linearAddr160List:
-         addr=self.addrMap[addr160]
-         addrList.append( [addr.chainIndex, addr160, addr] )
-
-      addrList.sort(key=lambda x: x[0])
-      return addrList
-
-   #############################################################################
-   def getAddrList(self):
-      """ Returns list of PyBtcAddress objects """
-      addrList = []
-      for addr160,addrObj in self.addrMap.items():
-         if addr160=='ROOT':
-            continue
-         # I assume these will be references, not copies
-         addrList.append( addrObj )
-      return addrList
-
-   #############################################################################
+   ## helpers
    def getLinearAddrList(self, withImported=True, withAddrPool=False):
       """
       Retrieves a list of addresses, by hash, in the order they 
@@ -772,9 +683,18 @@ class PyBtcWallet(object):
                addrList.append(addr)
       return addrList
 
-   #############################################################################
+   def getAddrListSortedByChainIndex(self, withRoot=False):
+      """ Returns Addr160 list """
+      addrList = []
+      for addr160 in self.linearAddr160List:
+         addr = self.addrMap[addr160]
+         addrList.append([addr.chainIndex, addr160, addr])
+
+      addrList.sort(key=lambda x: x[0])
+      return addrList
+
+   ####
    def pprint(self, indent='', allAddrInfo=True):
-      raise NotImplementedError("deprecated")
       print(indent + 'PyBtcWallet  :', self.walletId)
       print(indent + '   useEncrypt:', self.useEncryption)
       print(indent + '   watchOnly :', self.watchingOnly)
@@ -794,106 +714,7 @@ class PyBtcWallet(object):
             if allAddrInfo:
                addrObj.pprint(indent=indent)
 
-   #############################################################################
-   def isEqualTo(self, wlt2, debug=False):
-      isEqualTo = True
-      isEqualTo = isEqualTo and (self.walletId == wlt2.walletId and \
-         self.accountId == wlt2.accountId)
-      isEqualTo = isEqualTo and (self.labelName  == wlt2.labelName )
-      isEqualTo = isEqualTo and (self.labelDescr == wlt2.labelDescr)
-      try:
-         rootstr1 = binary_to_hex(self.addrMap['ROOT'].serialize())
-         rootstr2 = binary_to_hex(wlt2.addrMap['ROOT'].serialize())
-         isEqualTo = isEqualTo and (rootstr1 == rootstr2)
-         if debug:
-            print('')
-            print('RootAddrSelf:')
-            print(prettyHex(rootstr1, indent=' '*5))
-            print('RootAddrWlt2:')
-            print(prettyHex(rootstr2, indent=' '*5))
-            print('RootAddrDiff:', end=' ')
-            pprintDiff(rootstr1, rootstr2, indent=' '*5)
-
-         for addr160 in self.addrMap.keys():
-            addrstr1 = binary_to_hex(self.addrMap[addr160].serialize())
-            addrstr2 = binary_to_hex(wlt2.addrMap[addr160].serialize())
-            isEqualTo = isEqualTo and (addrstr1 == addrstr2)
-            if debug:
-               print('')
-               print('AddrSelf:', binary_to_hex(addr160), end=' ')
-               print(prettyHex(binary_to_hex(self.addrMap['ROOT'].serialize()), indent='     '))
-               print('AddrSelf:', binary_to_hex(addr160), end=' ')
-               print(prettyHex(binary_to_hex(wlt2.addrMap['ROOT'].serialize()), indent='     '))
-               print('AddrDiff:', end=' ')
-               pprintDiff(addrstr1, addrstr2, indent=' '*5)
-      except:
-         return False
-
-      return isEqualTo
-
-   #############################################################################
-   def toJSONMap(self):
-      outjson = {}
-      outjson['name']             = self.labelName
-      outjson['description']      = self.labelDescr
-      outjson['walletversion']    = getVersionString(PYBTCWALLET_VERSION)
-      outjson['balance']          = AmountToJSON(self.getBalance('Spend'))
-      outjson['keypoolsize']      = self.addrPoolSize
-      outjson['numaddrgen']       = len(self.addrMap)
-      outjson['highestusedindex'] = self._highestUsedChainIndex
-      outjson['watchingonly']     = self.watchingOnly
-      outjson['createdate']       = self.wltCreateDate
-      outjson['walletid']         = self.uniqueIDB58
-      outjson['isencrypted']      = self.useEncryption
-      outjson['islocked']         = self.isLocked if self.useEncryption else False
-      outjson['keylifetime']      = self.defaultKeyLifetime
-      return outjson
-
-   #############################################################################
-   def fromJSONMap(self, jsonMap, skipMagicCheck=False):
-      self.labelName   = jsonMap['name']
-      self.labelDescr  = jsonMap['description']
-      self.addrPoolSize  = jsonMap['keypoolsize']
-      self._highestUsedChainIndex  = jsonMap['highestusedindex']
-      self.watchingOnly  = jsonMap['watchingonly']
-      self.wltCreateDate  = jsonMap['createdate']
-      self.uniqueIDB58  = jsonMap['walletid']
-      jsonVer = hex_to_binary(jsonMap['walletversion'])
-
-      # Issue a warning if the versions don't match
-      if not jsonVer == getVersionString(PYBTCWALLET_VERSION):
-         LOGWARN('Unserializing wallet of different version')
-         LOGWARN('   Wallet Version: %d' % jsonVer)
-         LOGWARN('   Armory Version: %d' % UNSIGNED_TX_VERSION)
-
-   ###############################################################################
-   def getAddrTotalTxnCount(self, addrHash):
-      try:
-         addrObj = self.addrMap[addrHash]
-         return addrObj.getTxioCount()
-      except:
-         return 0
-
-   ###############################################################################
-   def getAddrDataFromDB(self):
-      result = self.bridgeWalletObj.getAddrCombinedList()
-
-      #update addr map
-      for addrProto in result.updatedAssets:
-         addrObj = PyBtcAddress()
-         addrObj.loadFromProto(addrProto)
-         self.addAddress(addrObj)
-
-      #update balances and txio count
-      for addrCombinedData in result.balances:
-         addr = addrCombinedData.scrAddr
-         if addr in self.addrMap:
-            addrObj = self.addrMap[addr]
-            addrObj.balance = Balances.fromProto(addrCombinedData.balances)
-         else:
-            print ("[getAddrDataFromDB] missing address " + addr.hex())
-
-   ###############################################################################
+   ####
    def getHistoryAsCSV(self, currentTop):
       file = open('%s.csv' % self.walletPath, 'wb')
 
@@ -928,68 +749,37 @@ class PyBtcWallet(object):
 
       file.close()
 
+   ####
+   def changePassphrase(self, isPriv, callbackId, callback):
+      self.bridgeWalletObj.changePassphrase(isPriv, callbackId, callback)
+
+   ####
+   def createAddressBook(self):
+      return self.bridgeWalletObj.createAddressBook()
+
+   ####
+   def setLabels(self, lshort, llong=''):
+      self.labelName = lshort
+      self.labelDescr = llong
+      self.bridgeWalletObj.setLabels(self.labelName, self.labelDescr)
+
+   ####
+   def delete(self):
+      return TheBridge.wltManager.deleteWallet(self.walletId)
+
    ###############################################################################
-   @CheckWalletRegistration
+   ## UI related
    def getHistoryPage(self, pageID):
       try:
          return self.cppWallet.getHistoryPage(pageID)
       except:
          raise Exception('pageID is out of range')
 
-   ###############################################################################
-   @CheckWalletRegistration
-   def doAfterScan(self):
-      actionsList = self.actionsToTakeAfterScan
-      self.actionsToTakeAfterScan = []
-
-      for calls in actionsList:
-         calls[0](*calls[1])
-
-   ###############################################################################
-   @CheckWalletRegistration
-   def sweepAfterRescan(self, addrList, main): 
-      #get a new address from the wallet to sweep the funds to
-      sweepToAddr = self.getNextUnusedAddress().getAddr160()
-
-      main.finishSweepScan(self, addrList, sweepToAddr)
-      return
-
-   ###############################################################################
-   @CheckWalletRegistration
-   def sweepAddressList(self, addrList, main):
-      self.actionsToTakeAfterScan.append(
-         [self.sweepAfterRescan, [addrList, main]])
-
-      addrVec = []
-      for addr in addrList:
-         addrVec.append(ADDRBYTE + addr.getAddr160())
-
-      _id = Cpp.SecureBinaryData().GenerateRandom(8).toHexStr()
-      main.oneTimeScanAction[_id] = self.doAfterScan()
-      TheBDM.bdv().registerAddrList(_id, addrList)
-
-   ###############################################################################
-   @CheckWalletRegistration
+   ####
    def disableWalletUI(self):
       self.isEnabled = False
 
-   ###############################################################################
-   def getAddrObjectForHash(self, hashVal):
-      assetIndex = self.cppWallet.getAssetIndexForAddr(hashVal)
-      if assetIndex == 2**32:
-         raise("unknown hash")
-
-      try:
-         addr160 = self.chainIndexMap[assetIndex]
-      except:
-         if assetIndex < -2:
-            importIndex = self.cppWallet.convertToImportIndex(assetIndex)
-            addr160 = self.linearAddr160List[importIndex]
-         else:
-            raise Exception("invalid address index")
-      return self.addrMap[addr160]
-
-   ###############################################################################
+   ####
    def returnFilteredAddrList(self, filterUse, filterType):
       addrList = []
       keepInUse = bool(filterUse != "Unused")
@@ -1001,7 +791,6 @@ class PyBtcWallet(object):
             addrList.append(addrObj)
       return addrList
 
-   ###############################################################################
    def getFilteredAddrCount(self, filterUse, filterType = None):
       count = 0
       keepInUse = bool(filterUse != "Unused")
@@ -1013,92 +802,12 @@ class PyBtcWallet(object):
             count += 1
       return count
 
-   ###############################################################################
-   def loadFromProto(self, payload):
-      self._walletId       = payload.walletId
-      self._accountId      = payload.accountId
-      self._dbId           = payload.dbId
-      self._settingsId     = computeSettingsId(self._walletId, self._accountId)
-      self.bridgeWalletObj = BridgeWalletWrapper(self._walletId, self._accountId)
-
-      self.labelName   = payload.label
-      self.labelDescr  = payload.desc
-
-      self.useEncryption = payload.usesEncryption
-      self.lastComputedChainIndex = payload.lookupCount
-      self._highestUsedChainIndex = payload.useCount
-      self.watchingOnly = payload.watchingOnly
-      self.addressTypes = payload.addressTypes
-      self.defaultAddressType = payload.defaultAddressType
-      self.kdfMemoryReq = payload.kdfMemReq * (1024**2)
-
-      #addrMap and chainIndexMap
-      for addr in payload.addressData:
-         addrObj = PyBtcAddress(self)
-         addrObj.loadFromProto(addr)
-         self.addAddress(addrObj)
-
-      #comments
-      for commentIt in payload.comments:
-         self.commentsMap[commentIt.key] = commentIt.val
-
-   #############################################################################
-   def fillAddressPool(self, numPool, progressId, callback=None):
-      """
-      This is only ever called by explicit user request. We therefor assume
-      the addresses are being restored (flagged as not new)
-      """
-      self.bridgeWalletObj.extendAddressPool(
-         numPool, False, progressId, callback)
-
-   #############################################################################
-   def register(self, isNew):
-      TheBridge.service.registerWallet(self.walletId, self.accountId, isNew)
-
-   #############################################################################
-   def createBackupString(self, callback,
-      passphrase: str=None, unlockHandler: callable=None):
-      return self.bridgeWalletObj.createBackupStringForWallet(callback,
-         passphrase, unlockHandler)
-
-   #############################################################################
-   def getAddressTypes(self):
-      return self.addressTypes
-
-   #############################################################################
-   def getDefaultAddressType(self):
-      return self.defaultAddressType
-
-   def getDefaultChangeType(self):
-      #TODO: add entry in cpp wallets for default change type
-      return self.defaultChangeType
-
-   #############################################################################
-   def setAddressTypeFor(self, addrObj, addrType):
-      if addrObj.addrType == addrType:
-         return
-
-      if addrType not in self.getAddressTypes():
-         raise Exception("ineligible address type")
-
-      protoAddr = self.bridgeWalletObj.setAddressTypeFor(
-         addrObj.assetId, addrType)
-      addrObj.loadFromProto(protoAddr)
-
-   #############################################################################
-   def initCoinSelectionInstance(self, height):
-      return self.bridgeWalletObj.initCoinSelectionInstance(height)
-
-   def createAddressBook(self):
-      return self.bridgeWalletObj.createAddressBook()
-
+   ####
    def getLedgerDelegateIdForScrAddr(self, scrAddr):
       return self.bridgeWalletObj.getLedgerDelegateIdForScrAddr(scrAddr)
 
-   def getKdfMemoryReqtBytes(self):
-      return self.kdfMemoryReq
-
    #############################################################################
+   ## properties
    @property
    def walletId(self):
       if not self._walletId:
@@ -1123,8 +832,66 @@ class PyBtcWallet(object):
          raise Exception("missing settingsId!")
       return self._settingsId
 
+   ####
+   def getHighestUsedIndex(self):
+      """
+      This only retrieves the stored value, but it may not be correct if,
+      for instance, the wallet was just imported but has been used before.
+      """
+      return self._highestUsedChainIndex + 1
+
+   def getHighestComputedIndex(self):
+      """
+      This only retrieves the stored value, but it may not be correct if,
+      for instance, the wallet was just imported but has been used before.
+      """
+      return self.lastComputedChainIndex
+
+   ####
+   def getWalletVersion(self):
+      return (getVersionInt(self.version), getVersionString(self.version))
+
+   ####
+   def getKdfMemoryReqtBytes(self):
+      return self.kdfMemoryReq
+
+   def testKdfComputeTime(self, callback):
+      """
+      Experimentally determines the compute time required by this computer
+      to execute with the current key-derivation parameters. This may be
+      useful for when you transfer a wallet to a new computer that has
+      different speed/memory characteristic.
+      """
+      def handleReply(reply):
+         if reply.success == False:
+            LOGERROR("failed to test KDF")
+         else:
+            self.testedComputeTime = reply.wallet.getUnlockTime
+         callback(reply.success, self.testedComputeTime)
+      self.bridgeWalletObj.getUnlockTime(handleReply)
+
+   ####
+   def getWalletPath(self, nameSuffix=None):
+      raise Exception("fixme")
+      fpath = self.walletPath
+
+      if self.walletPath=='':
+         fpath = os.path.join(ARMORY_HOME_DIR, buildWltFileName(self.uniqueIDB58))
+
+      if not nameSuffix==None:
+         pieces = os.path.splitext(fpath)
+         if not pieces[0].endswith('_'):
+            fpath = pieces[0] + '_' + nameSuffix + pieces[1]
+         else:
+            fpath = pieces[0] + nameSuffix + pieces[1]
+      return fpath
+
+   ####
+   def getDisplayStr(self, pref="Wallet: "):
+      return '%s"%s" (%s)' % (pref, self.labelName, self.walletId)
+
    #############################################################################
-   ## settings ##
+   ## settings
    def setSetting(self, propName, value):
       wltPropName = f'Wallet_{self.settingsId}_{propName}'
       TheSettings.set(wltPropName, value)
@@ -1139,3 +906,39 @@ class PyBtcWallet(object):
          if not defaultValue=='':
             self.setSetting(propName, defaultValue)
          return defaultValue
+
+   ###############################################################################
+   ## misc/deprecated/regressions
+   def doAfterScan(self):
+      actionsList = self.actionsToTakeAfterScan
+      self.actionsToTakeAfterScan = []
+
+      for calls in actionsList:
+         calls[0](*calls[1])
+
+   def sweepAfterRescan(self, addrList, main): 
+      #get a new address from the wallet to sweep the funds to
+      sweepToAddr = self.getNextUnusedAddress().getAddr160()
+
+      main.finishSweepScan(self, addrList, sweepToAddr)
+      return
+
+   def sweepAddressList(self, addrList, main):
+      self.actionsToTakeAfterScan.append(
+         [self.sweepAfterRescan, [addrList, main]])
+
+      addrVec = []
+      for addr in addrList:
+         addrVec.append(ADDRBYTE + addr.getAddr160())
+
+      _id = Cpp.SecureBinaryData().GenerateRandom(8).toHexStr()
+      main.oneTimeScanAction[_id] = self.doAfterScan()
+      TheBDM.bdv().registerAddrList(_id, addrList)
+
+   ####
+   def isWltSigningAnyLockbox(self, lockboxList):
+      for lockbox in lockboxList:
+         for addr160 in lockbox.a160List:
+            if addr160 in self.addrMap:
+               return True
+      return False
