+++ qtdialogs/DlgBackupCenter.py
@@ -14,19 +14,22 @@ from qtpy import QtCore, QtGui, QtWidgets
 
 from armoryengine.ArmoryUtils import toUnicode, USE_TESTNET, \
    USE_REGTEST, LOGEXCEPT, LOGINFO, LOGERROR
-from armoryengine.AddressUtils import binary_to_base58, encodePrivKeyBase58, \
-   hash160_to_addrStr
+from armoryengine.AddressUtils import binary_to_base58, \
+   encodePrivKeyBase58, hash160_to_addrStr
 from armorycolors import htmlColor, Colors
 
 from ui.WalletFrames import WalletBackupFrame
 from ui.QrCodeMatrix import CreateQRMatrix
+from ui.QtExecuteSignal import TheSignalExecution
 
 from qtdialogs.qtdefines import makeHorizFrame, QRichLabel, \
    makeVertFrame, QImageLabel, HLINE, GETFONT, STYLE_RAISED, tightSizeStr, \
-   setLayoutStretch, STRETCH, createToolTipWidget
+   setLayoutStretch, STRETCH, createToolTipWidget, MSGBOX
 from qtdialogs.ArmoryDialog import ArmoryDialog
 from qtdialogs.DlgUnlockWallet import UnlockWalletHandler
-from qtdialogs.DlgRestore import OpenPaperBackupDialog, getBackupTypeString
+from qtdialogs.DlgRestore import getBackupTypeString, \
+   DlgRestoreSingle, DlgRestoreFragged
+from qtdialogs.MsgBoxCustom import MsgBoxCustom
 
 ################################################################################
 class DlgBackupCenter(ArmoryDialog):
@@ -53,8 +56,6 @@ class DlgBackupCenter(ArmoryDialog):
 
 ################################################################################
 class DlgSimpleBackup(ArmoryDialog):
-
-   #############################################################################
    def __init__(self, parent, main, wlt):
       super(DlgSimpleBackup, self).__init__(parent, main)
       self.wlt = wlt
@@ -1364,3 +1365,57 @@ class DlgFragBackup(ArmoryDialog):
    def reject(self):
       self.destroyEverything()
       super(DlgFragBackup, self).reject()
+
+################################################################################
+def OpenPaperBackupDialog(backupType, parent, main, wlt, passphrase: str=None):
+   if passphrase:
+      #cleanup passphrase after 20sec
+      def cleanupPassphrase():
+         passphrase = None
+      TheSignalExecution.callLater(20, cleanupPassphrase)
+
+   result = True
+   verifyText = ''
+   if backupType == 'Single':
+      result = DlgPrintBackup(parent, main, wlt, passphrase=passphrase).exec_()
+      verifyText = parent.tr(
+         u'If the backup was printed with SecurePrint\u200b\u2122, please '
+         u'make sure you wrote the SecurePrint\u200b\u2122 code on the '
+         'printed sheet of paper. Note that the code <b><u>is</u></b> '
+         'case-sensitive!')
+   elif backupType == 'Frag':
+      result = DlgFragBackup(parent, main, wlt).exec_()
+      verifyText = parent.tr(
+         u'If the backup was created with SecurePrint\u200b\u2122, please '
+         u'make sure you wrote the SecurePrint\u200b\u2122 code on each '
+         'fragment (or stored with each file fragment). The code is the '
+         'same for all fragments.')
+
+   if result:
+      doTest = MsgBoxCustom(MSGBOX.Warning, parent.tr('Verify Your Backup!'),
+         parent.tr(
+            '<b><u>Verify your backup!</u></b> '
+            '<br><br>'
+            'If you just made a backup, make sure that it is correct! '
+            'The following steps are recommended to verify its integrity: '
+            '<br>'
+            '<ul>'
+            '<li>Verify each line of the backup data contains <b>9 columns</b> '
+            'of <b>4 letters each</b> (excluding any "ID" lines).</li> '
+            '<li>%s</li>'
+            '<li>Use Armory\'s backup tester to test the backup before you '
+            'physiclly secure it.</li> '
+            '</ul>'
+            '<br>'
+            'Armory has a backup tester that uses the exact same '
+            'process as restoring your wallet, but stops before it writes any '
+            'data to disk.  Would you like to test your backup now? '
+            % verifyText),
+         yesStr="Test Backup", noStr="Cancel")
+
+      if doTest:
+         if backupType == 'Single':
+            DlgRestoreSingle(parent, main, True, wlt.walletId).exec_()
+         elif backupType == 'Frag':
+            DlgRestoreFragged(parent, main, True, wlt.walletId).exec_()
+   return result
