+++ qtdialogs/DlgReplaceWallet.py
@@ -4,27 +4,22 @@
 # Distributed under the GNU Affero General Public License (AGPL v3)          #
 # See LICENSE or http://www.gnu.org/licenses/agpl.html                       #
 #                                                                            #
-# Copyright (C) 2016-2024, goatpig                                           #
+# Copyright (C) 2016-2025, goatpig                                           #
 #  Distributed under the MIT license                                         #
 #  See LICENSE-MIT or https://opensource.org/licenses/MIT                    #
 #                                                                            #
 ##############################################################################
 
-import os
-
 from qtpy import QtWidgets
-
-from armoryengine.ArmoryUtils import LOGEXCEPT, RightNowStr
-
 from qtdialogs.ArmoryDialog import ArmoryDialog
-from qtdialogs.DlgProgress import DlgProgress
+
+REPLACE_MERGE     = 1
+REPLACE_OVERWRITE = 2
 
 ################################################################################
 class DlgReplaceWallet(ArmoryDialog):
-
-   #############################################################################
    def __init__(self, WalletID, parent, main):
-      super(DlgReplaceWallet, self).__init__(parent, main)
+      super().__init__(parent, main)
 
       lblDesc = QtWidgets.QLabel(self.tr(
          '<b>You already have this wallet loaded!</b><br>'
@@ -43,11 +38,10 @@ class DlgReplaceWallet(ArmoryDialog):
       self.btnSaveMeta = QtWidgets.QPushButton(self.tr('Merge'))
 
       self.btnAbort.clicked.connect(self.reject)
-      self.btnReplace.clicked.connect(self.Replace)
-      self.btnSaveMeta.clicked.connect(self.SaveMeta)
+      self.btnReplace.clicked.connect(self.replace)
+      self.btnSaveMeta.clicked.connect(self.saveMeta)
 
       layoutDlg = QtWidgets.QGridLayout()
-
       layoutDlg.addWidget(lblDesc,          0, 0, 4, 4)
       layoutDlg.addWidget(self.btnAbort,    4, 0, 1, 1)
       layoutDlg.addWidget(self.btnSaveMeta, 4, 1, 1, 1)
@@ -57,11 +51,11 @@ class DlgReplaceWallet(ArmoryDialog):
       self.setWindowTitle('Wallet already exists')
 
    #########
-   def Replace(self):
-      self.output = 1
+   def replace(self):
+      self.output = REPLACE_OVERWRITE
       self.accept()
 
    #########
-   def SaveMeta(self):
-      self.output = 2
+   def saveMeta(self):
+      self.output = REPLACE_MERGE
       self.accept()
