+++ qtdialogs/DlgRestore.py
@@ -22,7 +22,7 @@ from ui.QtExecuteSignal import TheSignalExecution
 
 from qtdialogs.ArmoryDialog import ArmoryDialog
 from qtdialogs.DlgChangePassphrase import DlgChangePassphrase
-from qtdialogs.DlgReplaceWallet import DlgReplaceWallet
+from qtdialogs import DlgReplaceWallet as ReplaceWallet
 from qtdialogs.MsgBoxCustom import MsgBoxCustom
 from qtdialogs.qtdefines import HLINE, QRichLabel, STRETCH, STYLE_RAISED, \
    makeHorizFrame, makeVertFrame, MSGBOX, GETFONT, tightSizeStr, \
@@ -327,20 +327,21 @@ class DlgRestoreSingle(ArmoryDialog, ServerPush):
             #check if this wallet is already loaded
             dlgOwnWlt = None
             if self.main.wallets.hasWallet(newWltID):
-               dlgOwnWlt = DlgReplaceWallet(newWltID, self.parent, self.main)
-               if (dlgOwnWlt.exec_()):
-                  #unload existing wallet
-                  self.main.removeWalletFromApplication(newWltID)
-                  if dlgOwnWlt.output == 2:
+               dlgOwnWlt = ReplaceWallet.DlgReplaceWallet(
+                  newWltID, self.parent, self.main)
+               if dlgOwnWlt.exec_():
+                  if dlgOwnWlt.output == ReplaceWallet.REPLACE_MERGE:
                      replyToBridge.restore = 'merge'
-                  else:
+                  elif dlgOwnWlt.output == ReplaceWallet.REPLACE_OVERWRITE:
                      replyToBridge.restore = 'overwrite'
+                  self.main.removeWalletFromApplication(newWltID)
                else:
                   replyToBridge.success = False
 
          if replyToBridge.success == False:
             self.reject()
-         self.reply()
+         else:
+            self.reply()
 
       elif which == 'checksumError':
          checksums = restorePayload.checksumError
@@ -1531,60 +1532,6 @@ class DlgEnterSecurePrintCode(ArmoryDialog):
       if not checkSecurePrintCode(self, SECPRINT, securePrintCode):
          return
       self.accept()
-
-################################################################################
-def OpenPaperBackupDialog(backupType, parent, main, wlt, passphrase: str=None):
-   if passphrase:
-      #cleanup passphrase after 20sec
-      def cleanupPassphrase():
-         passphrase = None
-      TheSignalExecution.callLater(20, cleanupPassphrase)
-
-   result = True
-   verifyText = ''
-   if backupType == 'Single':
-      from qtdialogs.DlgBackupCenter import DlgPrintBackup
-      result = DlgPrintBackup(parent, main, wlt, passphrase=passphrase).exec_()
-      verifyText = parent.tr(
-         u'If the backup was printed with SecurePrint\u200b\u2122, please '
-         u'make sure you wrote the SecurePrint\u200b\u2122 code on the '
-         'printed sheet of paper. Note that the code <b><u>is</u></b> '
-         'case-sensitive!')
-   elif backupType == 'Frag':
-      result = DlgFragBackup(parent, main, wlt).exec_()
-      verifyText = parent.tr(
-         u'If the backup was created with SecurePrint\u200b\u2122, please '
-         u'make sure you wrote the SecurePrint\u200b\u2122 code on each '
-         'fragment (or stored with each file fragment). The code is the '
-         'same for all fragments.')
-
-   doTest = MsgBoxCustom(MSGBOX.Warning, parent.tr('Verify Your Backup!'),
-      parent.tr(
-         '<b><u>Verify your backup!</u></b> '
-         '<br><br>'
-         'If you just made a backup, make sure that it is correct! '
-         'The following steps are recommended to verify its integrity: '
-         '<br>'
-         '<ul>'
-         '<li>Verify each line of the backup data contains <b>9 columns</b> '
-         'of <b>4 letters each</b> (excluding any "ID" lines).</li> '
-         '<li>%s</li>'
-         '<li>Use Armory\'s backup tester to test the backup before you '
-         'physiclly secure it.</li> '
-         '</ul>'
-         '<br>'
-         'Armory has a backup tester that uses the exact same '
-         'process as restoring your wallet, but stops before it writes any '
-         'data to disk.  Would you like to test your backup now? '
-         % verifyText),
-      yesStr="Test Backup", noStr="Cancel")
-
-   if doTest:
-      if backupType == 'Single':
-         DlgRestoreSingle(parent, main, True, wlt.walletId).exec_()
-      elif backupType == 'Frag':
-         DlgRestoreFragged(parent, main, True, wlt.walletId).exec_()
-   return result
 
 ################################################################################
 def verifyRecoveryTestID(parent, computedWltID, expectedWltID=None):
