+++ qtdialogs/qtdialogs.py
@@ -1210,279 +1210,6 @@ class DlgDuplicateAddr(ArmoryDialog):
 
       self.setWindowTitle(self.tr('Duplicate Addresses'))
 
-
-################################################################################
-class DlgRemoveWallet(ArmoryDialog):
-   def __init__(self, wlt, parent=None, main=None):
-      super(DlgRemoveWallet, self).__init__(parent, main)
-
-      wltID = wlt.uniqueIDB58
-      wltName = wlt.labelName
-      wltDescr = wlt.labelDescr
-      lblWarning = QtWidgets.QLabel(self.tr('<b>!!! WARNING !!!</b>\n\n'))
-      lblWarning.setTextFormat(QtCore.Qt.RichText)
-      lblWarning.setAlignment(QtCore.Qt.AlignHCenter)
-
-      lblWarning2 = QtWidgets.QLabel(self.tr('<i>You have requested that the following wallet '
-                            'be removed from Armory:</i>'))
-      lblWarning.setTextFormat(QtCore.Qt.RichText)
-      lblWarning.setWordWrap(True)
-      lblWarning.setAlignment(QtCore.Qt.AlignHCenter)
-
-      lbls = []
-      lbls.append([])
-      lbls[0].append(QtWidgets.QLabel(self.tr('Wallet Unique ID:')))
-      lbls[0].append(QtWidgets.QLabel(wltID))
-      lbls.append([])
-      lbls[1].append(QtWidgets.QLabel(self.tr('Wallet Name:')))
-      lbls[1].append(QtWidgets.QLabel(wlt.labelName))
-      lbls.append([])
-      lbls[2].append(QtWidgets.QLabel(self.tr('Description:')))
-      lbls[2].append(QtWidgets.QLabel(wlt.labelDescr))
-      lbls[2][-1].setWordWrap(True)
-
-
-      # TODO:  This should not *ever* require a blockchain scan, because all
-      #        current wallets should already be registered and up-to-date.
-      #        But I should verify that this is actually the case.
-      wltEmpty = True
-      if TheBDM.getState() == BDM_BLOCKCHAIN_READY:
-         # Removed this line of code because it's part of the old BDM paradigm.
-         # Leaving this comment here in case it needs to be replaced by anything
-         # wlt.syncWithBlockchainLite()
-         bal = wlt.getBalance('Full')
-         lbls.append([])
-         lbls[3].append(QtWidgets.QLabel(self.tr('Current Balance (w/ unconfirmed):')))
-         if bal > 0:
-            lbls[3].append(QtWidgets.QLabel('<font color="red"><b>' + coin2str(bal, maxZeros=1).strip() + ' BTC</b></font>'))
-            lbls[3][-1].setTextFormat(QtCore.Qt.RichText)
-            wltEmpty = False
-         else:
-            lbls[3].append(QtWidgets.QLabel(coin2str(bal, maxZeros=1) + ' BTC'))
-
-
-      # Add two WARNING images on either side of dialog
-      lblWarnImg = QtWidgets.QLabel()
-      lblWarnImg.setPixmap(QtGui.QPixmap('./img/MsgBox_warning48.png'))
-      lblWarnImg.setAlignment(QtCore.Qt.AlignHCenter | QtCore.Qt.AlignVCenter)
-      lblWarnImg2 = QtWidgets.QLabel()
-      lblWarnImg2.setPixmap(QtGui.QPixmap('./img/MsgBox_warning48.png'))
-      lblWarnImg2.setAlignment(QtCore.Qt.AlignHCenter | QtCore.Qt.AlignVCenter)
-
-      # Add the warning text and images to the top of the dialog
-      layout = QtWidgets.QGridLayout()
-      layout.addWidget(lblWarning, 0, 1, 1, 1)
-      layout.addWidget(lblWarning2, 1, 1, 1, 1)
-      layout.addWidget(lblWarnImg, 0, 0, 2, 1)
-      layout.addWidget(lblWarnImg2, 0, 2, 2, 1)
-
-      frmInfo = QtWidgets.QFrame()
-      frmInfo.setFrameStyle(QtWidgets.QFrame.Box | QtWidgets.QFrame.Plain)
-      frmInfo.setLineWidth(3)
-      frmInfoLayout = QtWidgets.QGridLayout()
-      for i in range(len(lbls)):
-         lbls[i][0].setText('<b>' + lbls[i][0].text() + '</b>')
-         lbls[i][0].setTextFormat(QtCore.Qt.RichText)
-         frmInfoLayout.addWidget(lbls[i][0], i, 0)
-         frmInfoLayout.addWidget(lbls[i][1], i, 1, 1, 2)
-
-      frmInfo.setLayout(frmInfoLayout)
-      layout.addWidget(frmInfo, 2, 0, 2, 3)
-      hasWarningRow = False
-      if not wlt.watchingOnly:
-         if not wltEmpty:
-            lbl = QRichLabel(self.tr('<b>WALLET IS NOT EMPTY.  Only delete this wallet if you '
-                             'have a backup on paper or saved to a another location '
-                             'outside your settings directory.</b>'))
-            hasWarningRow = True
-         elif wlt.isWltSigningAnyLockbox(self.main.allLockboxes):
-            lbl = QRichLabel(self.tr('<b>WALLET IS PART OF A LOCKBOX.  Only delete this wallet if you '
-                             'have a backup on paper or saved to a another location '
-                             'outside your settings directory.</b>'))
-            hasWarningRow = True
-         if hasWarningRow:
-            lbls.append(lbl)
-            layout.addWidget(lbl, 4, 0, 1, 3)
-
-      self.radioDelete = QtWidgets.QRadioButton(self.tr('Permanently delete this wallet'))
-      self.radioWatch = QtWidgets.QRadioButton(self.tr('Delete private keys only, make watching-only'))
-
-      # Make sure that there can only be one selection
-      btngrp = QtWidgets.QButtonGroup(self)
-      btngrp.addButton(self.radioDelete)
-      if not self.main.usermode == USERMODE.Standard:
-         btngrp.addButton(self.radioWatch)
-      btngrp.setExclusive(True)
-
-      ttipDelete = self.main.createToolTipWidget(self.tr(
-         'This will delete the wallet file, removing '
-         'all its private keys from your settings directory. '
-         'If you intend to keep using addresses from this '
-         'wallet, do not select this option unless the wallet '
-         'is backed up elsewhere.'))
-      ttipWatch = self.main.createToolTipWidget(self.tr(
-         'This will delete the private keys from your wallet, '
-         'leaving you with a watching-only wallet, which can be '
-         'used to generate addresses and monitor incoming '
-         'payments.  This option would be used if you created '
-         'the wallet on this computer <i>in order to transfer '
-         'it to a different computer or device and want to '
-         'remove the private data from this system for security.</i>'))
-
-
-      self.chkPrintBackup = QtWidgets.QCheckBox(self.tr(
-         'Print a paper backup of this wallet before deleting'))
-
-      if wlt.watchingOnly:
-         ttipDelete = self.main.createToolTipWidget(self.tr(
-            'This will delete the wallet file from your system. '
-            'Since this is a watching-only wallet, no private keys '
-            'will be deleted.'))
-         ttipWatch = self.main.createToolTipWidget(self.tr(
-            'This wallet is already a watching-only wallet so this option '
-            'is pointless'))
-         self.radioWatch.setEnabled(False)
-         self.chkPrintBackup.setEnabled(False)
-
-
-      self.frm = []
-
-      rdoFrm = QtWidgets.QFrame()
-      rdoFrm.setFrameStyle(STYLE_RAISED)
-      rdoLayout = QtWidgets.QGridLayout()
-
-      startRow = 0
-      for rdo, ttip in [(self.radioDelete, ttipDelete), \
-                       (self.radioWatch, ttipWatch)]:
-         self.frm.append(QtWidgets.QFrame())
-         # self.frm[-1].setFrameStyle(STYLE_SUNKEN)
-         self.frm[-1].setFrameStyle(QtWidgets.QFrame.NoFrame)
-         frmLayout = QtWidgets.QHBoxLayout()
-         frmLayout.addWidget(rdo)
-         ttip.setAlignment(QtCore.Qt.AlignLeft | QtCore.Qt.AlignVCenter)
-         frmLayout.addWidget(ttip)
-         frmLayout.addStretch()
-         self.frm[-1].setLayout(frmLayout)
-         rdoLayout.addWidget(self.frm[-1], startRow, 0, 1, 3)
-         startRow += 1
-
-
-      self.radioDelete.setChecked(True)
-      rdoFrm.setLayout(rdoLayout)
-
-      startRow = 6 if not hasWarningRow else 5
-      layout.addWidget(rdoFrm, startRow, 0, 1, 3)
-
-      if wlt.watchingOnly:
-         self.frm[-1].setVisible(False)
-
-
-      printTtip = self.main.createToolTipWidget(self.tr(
-         'If this box is checked, you will have the ability to print off an '
-         'unencrypted version of your wallet before it is deleted.  <b>If '
-         'printing is unsuccessful, please press *CANCEL* on the print dialog '
-         'to prevent the delete operation from continuing</b>'))
-      printFrm = makeLayoutFrame(HORIZONTAL, [self.chkPrintBackup, \
-                                              printTtip, \
-                                              'Stretch'])
-      startRow += 1
-      layout.addWidget(printFrm, startRow, 0, 1, 3)
-
-      if wlt.watchingOnly:
-         printFrm.setVisible(False)
-
-
-      rmWalletSlot = lambda: self.removeWallet(wlt)
-
-      startRow += 1
-      self.btnDelete = QtWidgets.QPushButton(self.tr("Delete"))
-      self.btnCancel = QtWidgets.QPushButton(self.tr("Cancel"))
-      self.connect(self.btnDelete, SIGNAL(CLICKED), rmWalletSlot)
-      self.connect(self.btnCancel, SIGNAL(CLICKED), self.reject)
-      buttonBox = QtWidgets.QDialogButtonBox()
-      buttonBox.addButton(self.btnDelete, QtWidgets.QDialogButtonBox.AcceptRole)
-      buttonBox.addButton(self.btnCancel, QtWidgets.QDialogButtonBox.RejectRole)
-      layout.addWidget(buttonBox, startRow, 0, 1, 3)
-
-      self.setLayout(layout)
-      self.setWindowTitle(self.tr('Delete Wallet Options'))
-
-
-   def removeWallet(self, wlt):
-
-      # Open the print dialog.  If they hit cancel at any time, then
-      # we go back to the primary wallet-remove dialog without any other action
-      if self.chkPrintBackup.isChecked():
-         if not OpenPaperBackupDialog('Single', self, self.main, wlt, \
-                                                self.tr('Unlock Paper Backup')):
-            QtWidgets.QMessageBox.warning(self, self.tr('Operation Aborted'), self.tr(
-              'You requested a paper backup before deleting the wallet, but '
-              'clicked "Cancel" on the backup printing window.  So, the delete '
-              'operation was canceled as well.'), QtWidgets.QMessageBox.Ok)
-            return
-
-
-      # If they only want to exclude the wallet, we will add it to the excluded
-      # list and remove it from the application.  The wallet files will remain
-      # in the settings directory but will be ignored by Armory
-
-      wltID = wlt.uniqueIDB58
-      if wlt.watchingOnly:
-         reply = QtWidgets.QMessageBox.warning(self, self.tr('Confirm Delete'), \
-         self.tr('You are about to delete a watching-only wallet.  Are you sure '
-         'you want to do this?'), QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.Cancel)
-      elif self.radioDelete.isChecked():
-         reply = QtWidgets.QMessageBox.warning(self, self.tr('Are you absolutely sure?!?'), \
-         self.tr('Are you absolutely sure you want to permanently delete '
-         'this wallet?  Unless this wallet is saved on another device '
-         'you will permanently lose access to all the addresses in this '
-         'wallet.'), QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.Cancel)
-      elif self.radioWatch.isChecked():
-         reply = QtWidgets.QMessageBox.warning(self, self.tr('Are you absolutely sure?!?'), \
-         self.tr('<i>This will permanently delete the information you need to spend '
-         'funds from this wallet!</i>  You will only be able to receive '
-         'coins, but not spend them.  Only do this if you have another copy '
-         'of this wallet elsewhere, such as a paper backup or on an offline '
-         'computer with the full wallet.'), QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.Cancel)
-
-      if reply == QtWidgets.QMessageBox.Yes:
-
-         thepath = wlt.getWalletPath()
-         thepathBackup = wlt.getWalletPath('backup')
-
-         if self.radioWatch.isChecked():
-            LOGINFO('***Converting to watching-only wallet')
-            newWltPath = wlt.getWalletPath('WatchOnly')
-            wlt.forkOnlineWallet(newWltPath, wlt.labelName, wlt.labelDescr)
-            self.main.removeWalletFromApplication(wltID)
-
-            newWlt = PyBtcWallet().readWalletFile(newWltPath)
-            self.main.addWalletToApplication(newWlt, True)
-            # Removed this line of code because it's part of the old BDM paradigm.
-            # Leaving this comment here in case it needs to be replaced by anything
-            # newWlt.syncWithBlockchainLite()
-
-            os.remove(thepath)
-            os.remove(thepathBackup)
-            self.main.statusBar().showMessage( \
-               self.tr('Wallet %s was replaced with a watching-only wallet.' % wltID), 10000)
-
-         elif self.radioDelete.isChecked():
-            LOGINFO('***Completely deleting wallet')
-            if not wlt.delete():
-               LOGERROR("failed to delete wallet")
-               raise Exception("failed to delete wallet")
-
-            self.main.removeWalletFromApplication(wltID)
-            self.main.statusBar().showMessage( \
-               self.tr('Wallet %s was deleted!' % wltID), 10000)
-
-         self.parent.accept()
-         self.accept()
-      else:
-         self.reject()
-
-
 ################################################################################
 class DlgRemoveAddress(ArmoryDialog):
    def __init__(self, wlt, addr160, parent=None, main=None):
@@ -2217,412 +1944,6 @@ class ArmoryPref(object):
       if len(self.description) > 0:
          toDraw.append([QRichLabel(self.description), row, 0, 1, 4])
          row += 1
-
-
-################################################################################
-class QRadioButtonBackupCtr(QtWidgets.QRadioButton):
-   def __init__(self, parent, txt, index):
-      super(QRadioButtonBackupCtr, self).__init__(txt)
-      self.parent = parent
-      self.index = index
-
-
-   def enterEvent(self, ev):
-      pass
-      # self.parent.setDispFrame(self.index)
-      # self.setStyleSheet('QtWidgets.QRadioButton { background-color : %s }' % \
-                                          # htmlColor('SlightBkgdDark'))
-
-   def leaveEvent(self, ev):
-      pass
-      # self.parent.setDispFrame(-1)
-      # self.setStyleSheet('QtWidgets.QRadioButton { background-color : %s }' % \
-                                          # htmlColor('Background'))
-
-
-################################################################################
-# Class that acts as a center where the user can decide what to do with the
-# watch-only wallet. The data can be displayed, printed, or saved to a file as a
-# wallet or as the watch-only data (i.e., root key & chain code).
-class DlgExpWOWltData(ArmoryDialog):
-   """
-   This dialog will be used to export a wallet's root public key and chain code.
-   """
-   def __init__(self, wlt, parent, main):
-      super(DlgExpWOWltData, self).__init__(parent, main)
-
-      # Save a copy of the wallet.
-      self.wlt = wlt
-      self.main = main
-
-      # Get the chain code and uncompressed public key of info from the wallet,
-      # along with other useful info.
-      wltRootIDConcat, pkccET16Lines = wlt.getRootPKCCBackupData(True)
-      wltIDB58 = wlt.uniqueIDB58
-
-      # Create the data export buttons.
-      expWltButton = QtWidgets.QPushButton(self.tr('Export Watching-Only Wallet File'))
-      clipboardBtn = QtWidgets.QPushButton(self.tr('Copy to clipboard'))
-      clipboardLbl = QRichLabel('', hAlign=QtCore.Qt.AlignHCenter)
-      expDataButton = QtWidgets.QPushButton(self.tr('Save to Text File'))
-      printWODataButton = QtWidgets.QPushButton(self.tr('Print Root Data'))
-
-
-      self.connect(expWltButton, SIGNAL(CLICKED), self.clickedExpWlt)
-      self.connect(expDataButton, SIGNAL(CLICKED), self.clickedExpData)
-      self.connect(printWODataButton, SIGNAL(CLICKED), \
-                   self.clickedPrintWOData)
-
-
-      # Let's put the window together.
-      layout = QtWidgets.QVBoxLayout()
-
-      self.dispText = self.tr(
-         'Watch-Only Root ID:<br><b>%s</b>'
-         '<br><br>'
-         'Watch-Only Root Data:' % wltRootIDConcat)
-      for j in pkccET16Lines:
-         self.dispText += '<br><b>%s</b>' % (j)
-
-      titleStr = self.tr('Watch-Only Wallet Export')
-
-      self.txtLongDescr = QtWidgets.QTextBrowser()
-      self.txtLongDescr.setFont(GETFONT('Fixed', 9))
-      self.txtLongDescr.setHtml(self.dispText)
-      w,h = tightSizeNChar(self.txtLongDescr, 20)
-      self.txtLongDescr.setMaximumHeight(int(9.5*h))
-
-      def clippy():
-         clipb = QtWidgets.QApplication.clipboard()
-         clipb.clear()
-         clipb.setText(str(self.txtLongDescr.toPlainText()))
-         clipboardLbl.setText(self.tr('<i>Copied!</i>'))
-
-      self.connect(clipboardBtn, SIGNAL('clicked()'), clippy)
-
-
-      lblDescr = QRichLabel(self.tr(
-         '<center><b><u><font size=4 color="%s">Export Watch-Only '
-         'Wallet: %s</font></u></b></center> '
-         '<br>'
-         'Use a watching-only wallet on an online computer to distribute '
-         'payment addresses, verify transactions and monitor balances, but '
-         'without the ability to move the funds.' % (htmlColor('TextBlue'), wlt.uniqueIDB58)))
-
-      lblTopHalf = QRichLabel(self.tr(
-         '<center><b><u>Entire Wallet File</u></b></center> '
-         '<br>'
-         '<i><b><font color="%s">(Recommended)</font></b></i> '
-         'An exact copy of your wallet file but without any of the private '
-         'signing keys. All existing comments and labels will be carried '
-         'with the file. Use this option if it is easy to transfer files '
-         'from this system to the target system.' % htmlColor('TextBlue')))
-
-      lblBotHalf = QRichLabel(self.tr(
-         '<center><b><u>Only Root Data</u></b></center> '
-         '<br>'
-         'Same as above, but only five lines of text that are easy to '
-         'print, email inline, or copy by hand.  Only produces the '
-         'wallet addresses.   No comments or labels are carried with '
-         'it.'))
-
-      btnDone = QtWidgets.QPushButton(self.tr('Done'))
-      self.connect(btnDone, SIGNAL('clicked()'), self.accept)
-
-
-      frmButtons = makeVertFrame([clipboardBtn,
-                                  expDataButton,
-                                  printWODataButton,
-                                  clipboardLbl,
-                                  'Stretch'])
-      layoutBottom = QtWidgets.QHBoxLayout()
-      layoutBottom.addWidget(frmButtons, 0)
-      layoutBottom.addItem(QtWidgets.QSpacerItem(5,5))
-      layoutBottom.addWidget(self.txtLongDescr, 1)
-      layoutBottom.setSpacing(5)
-
-
-      layout.addWidget(lblDescr)
-      layout.addItem(QtWidgets.QSpacerItem(10, 10))
-      layout.addWidget(HLINE())
-      layout.addWidget(lblTopHalf, 1)
-      layout.addWidget(makeHorizFrame(['Stretch', expWltButton, 'Stretch']))
-      layout.addItem(QtWidgets.QSpacerItem(20, 20))
-      layout.addWidget(HLINE())
-      layout.addWidget(lblBotHalf, 1)
-      layout.addLayout(layoutBottom)
-      layout.addItem(QtWidgets.QSpacerItem(20, 20))
-      layout.addWidget(HLINE())
-      layout.addWidget(makeHorizFrame(['Stretch', btnDone]))
-      layout.setSpacing(3)
-
-      self.setLayout(layout)
-      self.setMinimumWidth(600)
-
-      # TODO:  Dear god this is terrible, but for my life I cannot figure
-      #        out how to move the vbar, because you can't do it until
-      #        the dialog is drawn which doesn't happen til after __init__.
-      self.callLater(0.05, self.resizeEvent)
-
-      self.setWindowTitle(titleStr)
-
-
-   def resizeEvent(self, ev=None):
-      super(DlgExpWOWltData, self).resizeEvent(ev)
-      vbar = self.txtLongDescr.verticalScrollBar()
-      vbar.setValue(vbar.minimum())
-
-
-   # The function that is executed when the user wants to back up the full
-   # watch-only wallet to a file.
-   def clickedExpWlt(self):
-      currPath = self.wlt.walletPath
-      if not self.wlt.watchingOnly:
-         pieces = os.path.splitext(currPath)
-         currPath = pieces[0] + '_WatchOnly' + pieces[1]
-
-      saveLoc = self.main.getFileSave('Save Watching-Only Copy', \
-                                      defaultFilename=currPath)
-      if not saveLoc.endswith('.wallet'):
-         saveLoc += '.wallet'
-
-      if not self.wlt.watchingOnly:
-         self.wlt.forkOnlineWallet(saveLoc, self.wlt.labelName, \
-                                '(Watching-Only) ' + self.wlt.labelDescr)
-      else:
-         self.wlt.writeFreshWalletFile(saveLoc)
-
-
-
-   # The function that is executed when the user wants to save the watch-only
-   # data to a file.
-   def clickedExpData(self):
-      self.main.makeWalletCopy(self, self.wlt, 'PKCC', 'rootpubkey')
-
-
-   # The function that is executed when the user wants to print the watch-only
-   # data.
-   def clickedPrintWOData(self):
-      self.result = DlgWODataPrintBackup(self, self.main, self.wlt).exec_()
-
-
-################################################################################
-# Class that handles the printing of the watch-only wallet data. The formatting
-# is mostly the same as a normal paper backup. Note that neither fragmented
-# backups nor SecurePrint are used.
-class DlgWODataPrintBackup(ArmoryDialog):
-   """
-   Open up a "Make Paper Backup" dialog, so the user can print out a hard
-   copy of whatever data they need to recover their wallet should they lose
-   it.
-   """
-   def __init__(self, parent, main, wlt):
-      super(DlgWODataPrintBackup, self).__init__(parent, main)
-
-      self.wlt = wlt
-
-      # Create the scene and the view.
-      self.scene = SimplePrintableGraphicsScene(self, self.main)
-      self.view = QtWidgets.QGraphicsView()
-      self.view.setRenderHint(QtGui.QPainter.TextAntialiasing)
-      self.view.setScene(self.scene.getScene())
-
-      # Label displayed above the sheet to be printed.
-      lblDescr = QRichLabel(self.tr(
-         '<b><u>Print Watch-Only Wallet Root</u></b><br><br> '
-         'The lines below are sufficient to calculate public keys '
-         'for every private key ever produced by the full wallet. '
-         'Importing this data to an online computer is sufficient '
-         'to receive and verify transactions, and monitor balances, '
-         'but without the ability to spend the funds.'))
-      lblDescr.setContentsMargins(5, 5, 5, 5)
-      frmDescr = makeHorizFrame([lblDescr], STYLE_RAISED)
-
-      # Buttons shown below the sheet to be printed.
-      self.btnPrint = QtWidgets.QPushButton('&Print...')
-      self.btnPrint.setMinimumWidth(3 * tightSizeStr(self.btnPrint, 'Print...')[0])
-      self.btnCancel = QtWidgets.QPushButton('&Cancel')
-      self.connect(self.btnPrint, SIGNAL(CLICKED), self.print_)
-      self.connect(self.btnCancel, SIGNAL(CLICKED), self.reject)
-      frmButtons = makeHorizFrame([self.btnCancel, STRETCH, self.btnPrint])
-
-      # Draw the sheet for the first time.
-      self.redrawBackup()
-
-      # Lay out the dialog.
-      layout = QtWidgets.QVBoxLayout()
-      layout.addWidget(frmDescr)
-      layout.addWidget(self.view)
-      layout.addWidget(frmButtons)
-      setLayoutStretch(layout, 0, 1, 0)
-      self.setLayout(layout)
-      self.setWindowIcon(QtGui.QIcon('./img/printer_icon.png'))
-      self.setWindowTitle('Print Watch-Only Root')
-
-      # Apparently I can't programmatically scroll until after it's painted
-      def scrollTop():
-         vbar = self.view.verticalScrollBar()
-         vbar.setValue(vbar.minimum())
-      self.callLater(0.01, scrollTop)
-
-
-   # Class called to redraw the print "canvas" when the data changes.
-   def redrawBackup(self):
-      self.createPrintScene()
-      self.view.update()
-
-
-   # Class that handles the actual printing code.
-   def print_(self):
-      LOGINFO('Printing!')
-      self.printer = QPrinter(QPrinter.HighResolution)
-      self.printer.setPageSize(QPrinter.Letter)
-
-      if QPrintDialog(self.printer).exec_():
-         painter = QtGui.QPainter(self.printer)
-         painter.setRenderHint(QtGui.QPainter.TextAntialiasing)
-
-         self.createPrintScene()
-         self.scene.getScene().render(painter)
-         painter.end()
-         self.accept()
-
-
-   # Class that lays out the actual print "canvas" to be printed.
-   def createPrintScene(self):
-      # Do initial setup.
-      self.scene.gfxScene.clear()
-      self.scene.resetCursor()
-
-      # Draw the background paper?
-      pr = self.scene.pageRect()
-      self.scene.drawRect(pr.width(), pr.height(), edgeColor=None, \
-                          fillColor=QtGui.QColor(255, 255, 255))
-      self.scene.resetCursor()
-
-      INCH = self.scene.INCH
-      MARGIN = self.scene.MARGIN_PIXELS
-      wrap = 0.9 * self.scene.pageRect().width()
-
-      # Start drawing the page.
-      if USE_TESTNET or USE_REGTEST:
-         self.scene.drawPixmapFile('./img/armory_logo_green_h56.png')
-      else:
-         self.scene.drawPixmapFile('./img/armory_logo_h36.png')
-      self.scene.newLine()
-
-      warnMsg = self.tr(
-         '<b><font size=4><font color="#aa0000">WARNING:</font>  <u>This is not '
-         'a wallet backup!</u></font></b> '
-         '<br><br>Please make a regular digital or paper backup of your wallet '
-         'to keep it protected!  This data simply lets you '
-         'monitor the funds in this wallet but gives you no ability to move any '
-         'funds.')
-      self.scene.drawText(warnMsg, GETFONT('Var', 9), wrapWidth=wrap)
-
-      self.scene.newLine(extra_dy=20)
-      self.scene.drawHLine()
-      self.scene.newLine(extra_dy=20)
-
-      # Print the wallet info.
-      colRect, rowHgt = self.scene.drawColumn(['<b>Watch-Only Root Data</b>',
-                                               'Wallet ID:',
-                                               'Wallet Name:'])
-      self.scene.moveCursor(15, 0)
-      colRect, rowHgt = self.scene.drawColumn(['',
-                                               self.wlt.uniqueIDB58,
-                                               self.wlt.labelName])
-
-      self.scene.moveCursor(15, colRect.y() + colRect.height(), absolute=True)
-
-      # Display warning about unprotected key data.
-      self.scene.newLine(extra_dy=20)
-      self.scene.drawHLine()
-      self.scene.newLine(extra_dy=20)
-
-      # Draw the description of the data.
-      descrMsg = self.tr(
-         'The following five lines are sufficient to reproduce all public '
-         'keys matching the private keys produced by the full wallet.')
-      self.scene.drawText(descrMsg, GETFONT('var', 8), wrapWidth=wrap)
-      self.scene.newLine(extra_dy=10)
-
-      # Prepare the data.
-      self.wltRootIDConcat, self.pkccET16Lines = \
-                                            self.wlt.getRootPKCCBackupData(True)
-      Lines = []
-      Prefix = []
-      Prefix.append('Watch-Only Root ID:');  Lines.append(self.wltRootIDConcat)
-      Prefix.append('Watch-Only Root:');     Lines.append(self.pkccET16Lines[0])
-      Prefix.append('');                     Lines.append(self.pkccET16Lines[1])
-      Prefix.append('');                     Lines.append(self.pkccET16Lines[2])
-      Prefix.append('');                     Lines.append(self.pkccET16Lines[3])
-
-      # Draw the prefix data.
-      origX, origY = self.scene.getCursorXY()
-      self.scene.moveCursor(10, 0)
-      colRect, rowHgt = self.scene.drawColumn(['<b>' + l + '</b>' \
-                                               for l in Prefix])
-
-      # Draw the data.
-      nudgeDown = 2  # because the differing font size makes it look unaligned
-      self.scene.moveCursor(10, nudgeDown)
-      self.scene.drawColumn(Lines,
-                              font=GETFONT('Fixed', 8, bold=True), \
-                              rowHeight=rowHgt,
-                              useHtml=False)
-
-      # Draw the rectangle around the data.
-      self.scene.moveCursor(MARGIN, colRect.y() - 2, absolute=True)
-      width = self.scene.pageRect().width() - 2 * MARGIN
-      self.scene.drawRect(width, colRect.height() + 7, \
-                          edgeColor=QtGui.QColor(0, 0, 0), fillColor=None)
-
-      # Draw the QR-related text below the data.
-      self.scene.newLine(extra_dy=30)
-      self.scene.drawText(self.tr(
-         'The following QR code is for convenience only.  It contains the '
-         'exact same data as the five lines above.  If you copy this data '
-         'by hand, you can safely ignore this QR code.'), wrapWidth=4 * INCH)
-
-      # Draw the QR code.
-      self.scene.moveCursor(20, 0)
-      x, y = self.scene.getCursorXY()
-      edgeRgt = self.scene.pageRect().width() - MARGIN
-      edgeBot = self.scene.pageRect().height() - MARGIN
-      qrSize = max(1.5 * INCH, min(edgeRgt - x, edgeBot - y, 2.0 * INCH))
-      self.scene.drawQR('\n'.join(Lines), qrSize)
-      self.scene.newLine(extra_dy=25)
-
-      # Clear the data and create a vertical scroll bar.
-      Lines = None
-      vbar = self.view.verticalScrollBar()
-      vbar.setValue(vbar.minimum())
-      self.view.update()
-
-
-def checkSecurePrintCode(context, SECPRINT, securePrintCode):
-   result = True
-   try:
-      if len(securePrintCode) < 9:
-         QtWidgets.QMessageBox.critical(context, context.tr('Invalid Code'), context.trUtf8(
-            u'You didn\'t enter a full SecurePrint\u200b\u2122 code.  This '
-            'code is needed to decrypt your backup file.'), QtWidgets.QMessageBox.Ok)
-         result = False
-      elif not SECPRINT['FUNC_CHKPWD'](securePrintCode):
-         QtWidgets.QMessageBox.critical(context, context.trUtf8(u'Bad SecurePrint\u200b\u2122 Code'), context.trUtf8(
-            u'The SecurePrint\u200b\u2122 code you entered has an error '
-            'in it.  Note that the code is case-sensitive.  Please verify '
-            'you entered it correctly and try again.'), QtWidgets.QMessageBox.Ok)
-         result = False
-   except NonBase58CharacterError as e:
-      QtWidgets.QMessageBox.critical(context, context.trUtf8(u'Bad SecurePrint\u200b\u2122 Code'), context.trUtf8(
-         u'The SecurePrint\u200b\u2122 code you entered has unrecognized characters '
-         'in it.  %s Only the following characters are allowed: %s' % (e.message, BASE58CHARS)), QtWidgets.QMessageBox.Ok)
-      result = False
-   return result
-
-
 
 ##########################################################################
 class DlgShowTestResults(ArmoryDialog):
